anda

+++ doc_offsets.h
@@ -17,9 +17,9 @@ class Doc_offsets {
 public:
 	Doc_offsets(const char* nombre);
 	virtual ~Doc_offsets();
-	void add_cant_doc(int cant);
-	void add_num_doc(int doc);
-	void add_frec(int frec);
+	void add_cant_doc(size_t cant);
+	void add_num_doc(size_t doc);
+	void add_frec(size_t frec);
 	void add_offset(size_t offset);
 private:
 	Traductor* unTraductor;
